[Commons] Change some db-related fields type

+++ dto/common.go
@@ -1,9 +1,9 @@
 package dto
 
 type ListMetadata struct {
-	Count  int64 `json:"count"`
-	Offset int64 `json:"offset"`
-	Limit  int64 `json:"limit"`
+	Count  int   `json:"count"`
+	Offset int   `json:"offset"`
+	Limit  int   `json:"limit"`
 	Total  int64 `json:"total"`
 }
 
@@ -13,6 +13,6 @@ type ListResp struct {
 }
 
 type ListRequest struct {
-	Limit  int64
-	Offset int64
+	Limit  int
+	Offset int
 }
